chore(Select): Refactor folder structure
